refactor(backend): optimize handlers
