+++ apps/backend/src/handlers/index.ts
@@ -1,5 +1,3 @@
 export * from "./publication";
-export * from "./mirror";
 export * from "./momoka"
 export * from "./send";
-export * from "./collect";
