+++ apps/backend/src/handlers/publication.ts
@@ -1,115 +1,48 @@
-import { CommentFragment, PostFragment } from "@lens-protocol/client";
+import { AnyPublicationFragment } from "@lens-protocol/client";
 import {
   MessageContent,
   PublicationEmbed,
-  getMonitoredProfileIds,
-  lensHubInterface,
   getPublicationUrl,
   getPictureUrl,
-  getProfile,
-  numberToHex,
-  parseUri,
-  getPublicationById,
-  getProfileUrl,
+  hexToNumber,
 } from "../utils";
-import { ILog } from "../types";
-import { sendToDiscord } from "./send";
-import { Log } from "ethers";
+import { sendToDiscord } from ".";
 
 export const handlePublication = async (
-  type: "Post" | "Comment",
-  log: Log,
-  txHash: string
+  publication: AnyPublicationFragment
 ) => {
-  const decoded = lensHubInterface.parseLog(log as unknown as ILog);
-  const monitoredProfileIds = await getMonitoredProfileIds();
+  const profile = publication.by;
+  const publicationUrl = getPublicationUrl(publication.id);
 
-  if (!decoded || !monitoredProfileIds.includes(decoded.args[0].toString()))
-    return;
+  const type = publication.__typename!;
 
-  const [profileId, pubId, contentUri] = decoded!.args;
-  try {
-    const parsedUri = parseUri(contentUri);
-
-    const [profile, metadata] = await Promise.all([
-      getProfile(numberToHex(profileId)),
-      fetch(parsedUri).then((res) => res.json()),
-    ]);
-
-    if (!profile || !metadata) {
-      return new Error(
-        `Failed to fetch data; profileId: ${profileId}; tx: ${txHash}`
-      );
-    }
-
-    const publicationId = `${profile.id}-${numberToHex(pubId)}`;
-
-    let targetHandle: string | undefined = undefined;
-    if (type == "Comment") {
-      const pub = (await getPublicationById(pubId)) as CommentFragment;
-      targetHandle =
-        pub.commentOn?.by.handle?.localName ?? pub.commentOn?.by.id;
-    }
-
-    let content = MessageContent(
-      type + "ed",
-      getPublicationUrl(publicationId),
-      targetHandle
-    );
-    const embeds = PublicationEmbed({
-      id: publicationId,
-      metadata,
-      profile,
-      appId: metadata.appId,
-    });
-
-    const quotedPost = metadata.attributes.find(
-      // @ts-ignore
-      (attribute) => attribute.traitType == "quotedPublicationId"
-    );
-
-    if (quotedPost) {
-      let quotedPub = (await getPublicationById(
-        quotedPost.value
-      )) as PostFragment | null;
-
-      if (quotedPub) {
-        const quotedHandle = quotedPub.by.handle?.localName ?? quotedPub.by.id;
-        content = `[Quoted](${getPublicationUrl(
-          publicationId
-        )}) [@${quotedHandle}](${getProfileUrl(quotedHandle)})`;
-        embeds.push(
-          ...PublicationEmbed({
-            id: quotedPub.id,
-            metadata: quotedPub.metadata,
-            profile: quotedPub.by,
-          })
-        );
-      } else {
-        return new Error(
-          `Quoted publication not found (tx: ${txHash}; id: ${pubId}; quotedId: ${quotedPost.value})`
-        );
-      }
-    }
-
-    const payload = {
-      username:
-        profile.metadata?.displayName ??
-        profile.handle?.localName ??
-        profile.id,
-      avatar_url: getPictureUrl(profile),
-      content,
-      embeds,
-    };
-
-    await sendToDiscord({
-      profileId: profileId.toString(),
-      type,
-      payload,
-    });
-  } catch (error) {
-    return new Error(
-      `Error handling publication (tx: ${txHash}; id: ${pubId}): ${error}`
-    );
+  let posts;
+  if (publication.__typename == "Mirror") {
+    posts = [publication.mirrorOn];
+  } else if (publication.__typename == "Comment") {
+    posts = [publication.commentOn, publication];
+  } else if (publication.__typename == "Quote") {
+    posts = [publication, publication.quoteOn];
+  } else {
+    posts = [publication];
   }
+
+  let content = MessageContent(type, publicationUrl);
+
+  const embeds = posts.map((post) => PublicationEmbed(post)).flat();
+
+  const payload = {
+    username:
+      profile.metadata?.displayName ?? profile.handle?.fullHandle ?? profile.id,
+    avatar_url: getPictureUrl(profile),
+    content,
+    embeds,
+  };
+
+
+  await sendToDiscord({
+    profileId: hexToNumber(profile.id),
+    type,
+    payload,
+  });
 };
